fix: QR 코드 모달 스타일 수정
